Fix UnmanagedCallingConventions test reliability
The test needs the `CombineStrings` instane to be alive through the
duration of the function pointer call. The `GC.KeepAlive` call though
was on the returned `IntPtr`. That is essentially a no-op cause it just
boxes to `object`. It doesn't actaully track the underlying `delegate`
instance.

Changed the code to track the `delegate` and push the `KeepAlive` after
the function pointer invocation.

closes #48641