Add usings on paste (#48501)
Addresses #45853

Hook up paste command to also automatically add usings. Operation uses thread await dialog to pop up to a user if it is taking a while with a "cancel" button. The dialog will not always show, but follows the IUIThreadOperationContext scope rules

Edit:

This is disabled by default. The following followup issues were filed
#49442
#49443
#49444