Make compiler server logging explicit (#48557)
* Make compiler server logging explicit

This changes the compiler server logging infrastructure to be an
explicit resource vs. an ambient authority. This both cleans up our
implementation a bit but also allows us to take advantage of xUnit
output logging in our unit tests.

The `ITestOutputHelper` can be the implementation of an
`ICompilerServerLogger`. This means that when compiler server tests fail
we get the full benefit of our logging infrastructure in the test
output vs. having to just guess at what failed here.