Fix the ability to expand the list of analyzers in a reference
This was broken by 7fa2c964f8433b928c5ff966427ce3798724cdd2, where I
asserted that GetDiagnosticItems was only called when we already knew
we would have items. This was broken since right before we called the
method the collection would get created, which changed the answer
it would return. A small refactoring ensures we create the list and
fill it in at once. This also makes it clear we aren't passing a
non-trivial LINQ query across method boundaries anyways in a way that
might result in multiple enumerations.

Fixes https://github.com/dotnet/roslyn/issues/49524