Show generated files live in Solution Explorer
This updates the Solution Explorer to show the generators underneath
an analyzer node like we do diagnostics today, and under the generator
node we show individual files. I'm doing this as a two-stage hierarchy
for a few reasons:

1. Computing the list of generated files isn't cheap (we need to run
   the generators, after all), so that way you don't see them or pay
   overhead unless you really want to see them. Expanding an analyzer
   to see if it has configurable diagnostics doesn't pay any cost.
2. It's a bit easier to see which generators are generating which files,
   if you have multiple generators in the same assembly. I'm not sure
   if that matters or not in the real world, but if you had two
   generators in the same assembly and they generated two files with the
   same hint path, you have some easy way to differentiate.
3. It kept the code a bit simpler to maintain, although I now realize
   that I could have split the BaseDiagnosticAndGeneratorItemSource into
   two separate IAttachedCollectionSources.

We keep the list updated in the background if it's expanded. We refresh
very slowly (only after five seconds of no activity) after the initial
expansion to avoid unnecessary work; in practice a lot of generators
will produce a fairly stable list of hint names from one run to the
next, so most people won't notice.

You can double click an item to open the dynamic file view.