Fix crash where we try previewing a generated document; general cleanup
The core problem here was when we were trying to create a preview,
we were trying to fork the contents of the document with a SourceText
that came from the buffer; if the document is a source generated
document trying to fork it won't work, and we'd crash.

Upon further inspection this pattern of using PreviewWorkspace this way
was widespread; we'd fork a document with new content, and then call
PreviewWorkspace.OpenDocument() which would fetch the text back and then
open it up. This doesn't really make sense: OpenDocument updates the
text as a part of it's action, so the extra document forking is entirely
unnecessary if we just pass the container in directly when opening the
document. Furthermore some code in AbstractPreviewFactoryService was
also removing and adding documents for other reasons I can't explain,
but we can delete all of this to clean things up.

After that refactoring, the fix is to simply not open the source
generated document, which we do by checking in
PreviewWorkspace.OpenDocument for now. We don't support opening source
generated documents at this time, but once we do in the core Workspace
layer we'll enable that so we get classification support in previews.

Fixes https://github.com/dotnet/roslyn/issues/49755