Move IEditorFormattingService to the Features layer
Today, IEditorFormattingService lives in the EditorFeatures layer, which is an impediment to making the API public and to making our LSP depend only on the Features layer. This moves the service down to the Features layer. The service was almost entirely independent of the EditorFeatures layer today, with one exception: it used [Microsoft.VisualStudio.TextEditor.IIndentationManager.GetIndentation](https://docs.microsoft.com/en-us/dotnet/api/microsoft.visualstudio.text.editor.iindentationmanagerservice.getindentation?view=visualstudiosdk-2019#Microsoft_VisualStudio_Text_Editor_IIndentationManagerService_GetIndentation_Microsoft_VisualStudio_Text_ITextBuffer_System_Boolean_System_Boolean__System_Int32__System_Int32__), which of course cannot be directly moved to the Features layer. I extracted the extension method that called this into it's own workspace service, with a no-op implementation in the Workspace layer, and a VS-specific implementation in the VisualStudio layer that overrides the no-op version and calls the VS-specific API. The implementation is otherwise almost entirely untouched, just a couple of small housekeeping changes while I was moving the file.