Extract IFindDefinitionService into to the features layer (#50132)
Previously IGotoDefinitionService was used in LSP, for functionality that can entirely exist in the Features layer instead of the EditorFeatures layer. This is the first, non-breaking-change part that leaves the existing method as is, and tries to find an implementation of IFindDefinitionService first, falling back if not found. https://github.com/dotnet/roslyn/issues/50391 tracks removing the fallback when our partner teams have implemented the new interface.

Co-authored-by: CyrusNajmabadi <cyrus.najmabadi@gmail.com>