Remove ConditionalFact from MSBuildWorkspace tests
We absolutely allow MSBuildWorkspace to be used on 64-bit; it seems the
real reason this test was failing was because it didn't correctly
ensure that registers or the stack might accidentally root the Solution
object it was trying to assert: specifically, the assert about the
cache service being null would have resulted in the object we wanted
to assert is weakly held being placed as a temporary. On 64-bit it seems
the JIT has no reason to reuse that slot/register so the test would
fail.