Remove some very outdated comments
The comment about the interop DLL dates back to when Roslyn shipped
a native DLL to deal with some COM interop issues; that hasn't been true
for many years so the comment is unnecessary.

The comment for the workspace not being created before loading the
package is also not correct: we've made that safe for quite some time
now.