Correctly attach analyzer nodes to multi-targeting projects
We are responsible for filling out the contents under the Analyzers
node where we put the list of analyzers, generators, and information
about them. To do this we have to go from an IVsHierarchy and item
in the Solution Explorer to a Roslyn project. Our code was trying to
take the IVsHierarchy and item to get a project path and TFM, and then
trying to walk through our projects to find a project associated with
the same project path and TFM. This code however was written back when
there was a wrapper hierarchy for each project in a multi-targeted
project which hasn't been the case for some time.

The project system is now passing us what is roughly a TFM through
IWorkspaceProjectContext which we can use to match up and select the
correct Roslyn Project in the multitargeting case.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1275676