Include the ProjectId's GUID in the computation of the DocumentId
The equality fo a DocumentId is defined by comparing the GUID of the
DocumentId and the GUID of the ProjectId, so it's already the case that
if we have two generated documents with the same hint path we'll end
up with different DocumentIds. However, if anybody tries getting
'clever' with the GUID of the DocumentId (or are just inspecting it
in a debugger) it might get confusing. This ensures they're different.

Fixes https://github.com/dotnet/roslyn/issues/50714