Simplify state management in AbstractSnippetExpansionClient
This takes advantage of the prior commit to simplify some state
management. When we were using snippet functions to populate each
parameter, it was necessary to ensure some state "survives" when we are
moving from one overload to the next. This is because there's some
reentrancy: the process of moving to a new overload would dismiss the
prior snippet, but it had to still be avialable before the call to
insert the new snippet returned. Now that the snippet is fully
pre-generated, that's not a requirement anymore; instead we can just
always clear state when one snippet is dismissed, and set it afterwards
again once the snippet is re-inserted. This removes the necessity of
having the "_preserveSymbols" flag, as well as removing the requirement
to call specific Clear() at various points when we were dismissing
sessions.

The other simplification is the list of methods that was stored once a
session began was only used to grab one thing: the name of the method
group. So just replace that with a string.