Rebuild API shape (#52079)
* Rebuild API shape

This breaks up `BuildConstructor` into a simple hierarchy with both
CSharp and Visual Basic implementations. The name `CompilationFactory`
was chosen after a bit of offline conversation and mostly due to this
representing a factory of sorts. As with all other API names it will be
reviewed before making this API officially public / shipping.

Note: there is no intended functional change with this PR, it is purely
just refactoring the API shape around.

* Apply suggestions from code review

Co-authored-by: Rikki Gibson <rikkigibson@gmail.com>

* PR Feedback

* PR feedback

* PR feedback

Co-authored-by: Rikki Gibson <rikkigibson@gmail.com>