Delete AssertEx.Throws
We're not entirely sure why this was added; it might have been due
to lack of Assert.ThrowsAsync in whichever version of xUnit we were
using, or something else. In any case, it's unnecessary today and was
also broken: it would silently succeed if the passed function didn't
throw any exceptions at all.