Use CommonCompiler in Rebuild tests (#52208)
* Add file system abstraction

This adds a file system abstraction sufficient to hook all file open
events inside `CommonCompiler`. This is necessary to fully test the
rebuild scenarios that we need to add

* Generated code has access to ref name

* Tests for round tripping through CommonCompiler

* Fix debug plus round trip

* Fix failing tests

The determinism / round trip tests in emit were depending on a bug in
our test framework. The metadata writer only emits references if there
is a name associated with the reference. Before my PR all of our in
memory references lacked image names and hence they were omitted from
the PDB. Once I fixed that they started showing up and needed to be
accounted for in the logic.

* Fix a build error

* Respond to PR feedback

* Fix test failures

* PR feedback

* PR feedback

* Test failure

* Fixups

* PR feedback