Replace iff with "if and only if" (#52294)
* Replace iff with "if and only if"

Fixes dotnet/docs#21572

* replace "iff" with "if and only if" in other public `///` locations

PR review pointed to two other locations where "iff" was used in the /// comments. This commit replaces those with "if and only if" in those cases where the API is public, or protected and will be part of the documentation set.

There are other uses in comments, include `///` comments on internal (or Friend) APIs  and private methods. Those are unchanged.