Dynamically connect the source generated file to the workspace
We don't know up-front whether a generated file exists in the workspace,
and our APIs to grab a document from a buffer are intended to complete
quickly. If it turns out later the file isn't in the workspace anymore
we will fake it and add it back so that way features aren't surprised by
this. However, that means semantics in that file may be inaccurate.
Disconnecting the document means that once the file is gone, then we
won't be in this incorrect state for long.