Fix failure to process RoslynEventSource manifest
Due to a lack of NonEventAttribute on the compiler-generated method for
a lambda function, the dynamic manifest treated a method named
"<OnEventCommand>b__8_0" as an event and included the name in the
manifest without escaping. The lack of escaping rendered the entire
manifest invalid XML, so it could not be used for interpreting events in
PerfView.