Addressed PR Feedback
I restored IEditorFormattingService to the EditorServices layer, and renamed the moved Features layer service to IFormattingInteractionService based on offline feedback. I've also added the TS and F# ExternalAccess APIs so that they can move to the new version, and updated the locations that use IEditorFormattingService to prefer IFormattingInteractionService, with a fallback if it doesn't exist.