Reuse the prior compilation with generated trees if possible
The CompilationTracker, even if we were reusing existing trees from a
prior run, would always take the Compilation that didn't have the
generated syntax trees and would add the generated trees to produce the
final compilation. This change tries to make it smarter, so if the
generated trees are identical, we instead take the prior compilation
with generated trees and instead update that with the updates to
non-generated trees. This can help us avoid having to recompute the
declaration tables for the generated trees (which can often be large)
multiple times.