Delete IWaitIndicator in favor of the public platform types
This interface was a MEF abstraction that Roslyn used over the threaded
wait dialog in Visual Studio; since we didn't want to tie our code
directly to the VS interfaces, we introduced this abstraction so hosts
could provide their own implementation. This interface was always
internal, which of course meant that any host (like Visual Studio for
Mac) that wasn't directly in the Roslyn codebase had to have
InternalsVisibleTo access.

This abstraction was so good, in the intervening years the VS platform
needed a similar concept, and so the platform introduced
IUIThreadOperationExecutor, which was their name for basically the
same thing. But unlike ours (which was internal), theirs is public,
and already has implementations for our hosts. And anybody wanting to
host Roslyn already will have to define the editor version, so having
to also force hosts to define a second implementation is a waste of
everybody's time. Specifically this often creates a lot of confusion
around creating MEF containers when it's just not necessary anymore.

In this PR I'm moving Roslyn off of our internal interface but leaving
the implementation there; that's because we have some other users via
InternalsVisibleTo that are still using this directly. The methods
and types are now marked as Obsolete, and the users can simply
migrate to the editor APIs like we are too.

There is one extra thing our implementation also did which the platform
one won't, and that's talk to our global operation service which is
potentially used to pause long-running operations when there is a user-
blocking operation active. I'm skeptical that is providing much value,
because when we migrated to the editor's command handling, we migrated
most of our uses of IWaitIndicator to use the platform provided context,
and it seems everything was fine.

As a small note, our implementation of IWaitIndicator also implemented
a second copy of the interface that's defined in
Microsoft.VtsualStudio.Text.Internal. I believe the editor has wanted
us to stop doing that for a long time since they couldn't delete those
types, even if they themselves had stopped using it.

The editor also provides a default implementation of this interface so
we don't even have to provide a test one for unit tests, which is nice.