Improve Server Logging (#52572)
* Improve compiler server logging

The goal if this PR is to change the logging so that:

1. It's easier to identify the process writing the message. The prefix
for messages will be prefixed with either "VBCSCompiler" or "MSBuild
<process id>".
1. Add a `Guid` to a logical build request that is included in many
messages. This will make it easier to track requests as they travel
between the MSBuild and VBCSCompiler processes.
1. Ensure all failure states have an explicit log call

The motivation for this change is we are seeing the compiler server
crash in build correctness and source build legs. From the log it's hard
to see what is happening. The failure is essentially silent. This is an
attempt to get more data so we can see what is actually going on during
failure.

* More

* Capture dumps when the compiler crashes

* PR feedback