Enable on-demand file logging for diagnostics
Fixes #51521
Adds a new option to enable logging diagnostics information to a log file inside the temp folder, which the user can submit for any intellisense diagnostics/error list related issues. We are not investing in a full-fledged system here as we anticipate many architectural changes in the diagnostics system in Dev17. The support added here is a short term solution to enable getting additional logs.