Fix code cleanup registration for VB
This tried to use the VB content type name for what should have been
the VB project capability. To avoid confusion, the C# one is also
updated -- the C# content type and capability happen to be the same
string value, but avoiding use of ContentTypeNames makes it clear
AppliesToProject is not taking a content type, conceptually.