Implement binding and lowering for InterpolatedStringBuilder (#52422)
* Implement binding and lowering for InterpolatedStringBuilder

This implements initial binding and lowering for the InterpolatedStringBuilder, used when an interpolated string is converted directly to a `string`, and not to some other builder type. I've written it in such a way as to make future refactorings to support general builder types simple, but some of the steps (particularly lowering) are more simplistic for now. I did _not_ implement nullable analysis or definite assignment in this PR, prototype comments have been left to make sure that those are handled later. The spec for this feature is located at https://github.com/dotnet/csharplang/blob/main/proposals/improved-interpolated-strings.md, I will have a PR out later today or tomorrow to reflect the changes from LDM earlier today.