Correctly find enum members in case labels
When we changed the logic for finding declared symbols in
ca2505a08c8974dde3243bd9fdb6a159c3d83cd7, we now would see the case
label as declaring the symbol which we don't want. As a conservative
choice, we'll do the expanded logic only for operators where it was
originally intended.

Fixes https://github.com/dotnet/roslyn/issues/53269