Allow our InProc "OOP" provider to share TestGeneratorReferences
For testing OOP serialization in unit tests, we spin up an in-proc
"remote" provider which does all the serialization and synchronization
like we would for OOP, just without the actual separate process. One
problem for unit tests was this had the some of the same limitations
of our OOP sync. Specifically, we only supported synchronizing
AnalyzerFileReferences, since that was implemented by loading the
analyzer on the other side. For unit tests around source generator
support, we have a TestGeneratorReference helper which is an
AnalyzerReference that directly holds onto an instance of a generator.
Since a TestGeneratorReference isn't an AnalyzerFileReference, this
meant we couldn't test scenarios that had generators in our "remote"
provider we used in unit tests.

This change cheats a little bit in the OOP serialization layer,
and in tests allows us to "serialize" a TestGeneratorReference to
the remote workspace so the generator is running there too. In our
unit tests we create an entirely separate MEF composition for the
remote workspace to ensure isolation, but we'll cheat and share a single
map of GUID to TestSourceGenerator between the two MEF compositions.
"Serializing" a TestGeneratorReference is then adding something into
that map, and the other side can read back out.

This is a part of https://github.com/dotnet/roslyn/issues/50494.