Don't attempt to edit source generated files during rename
We don't let a user rename symbols that are defined in source generated
files, but in some scenarios we may cascade to them. This prevents us
from updating those cascaded definitions either. The assumption in such
a change is that the names in generated files are being generated from
some of the original definitions that are being renamed, and so
everything will refresh.

Fixes https://github.com/dotnet/roslyn/issues/51537

# Conflicts:
#	src/EditorFeatures/Test2/Rename/RenameEngineResult.vb