Update contrib documentation (#53504)
The PR #53138 moved us to using .NET 6 Preview 3 but did not update our
documentation. Historically moving to new .NET SDK versions was manual
process and we "knew" to update the docs when we updated the version.
Now though it's flowing automatically via arcade and we missed it.

Short term this will get the docs in line with the requirements. Longer
term though working on a way to make sure automation doesn't silently
break our docs here.