Don't hold onto some SourceTexts if multiple changes happen
If we are processing edits through TryApplyChanges for closed files,
we'll make a series of changes in rapid succession: we'll have one
change that may open the file in an invisible editor, then we'll
edit the file, and then close again. The Solution object would then
be holding onto the old Compilation with several changes queued, which
would root the intermediate states that still pointed the editor
contents and their change lists which can be large. Once a Compilation
was requested for this new file, we'd process all the changes and
free the memory, but if the file is closed we might not get to that
right away.